+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Latest:
  - 
  - 
  - Add XChaCha20 CipherKey.
+ - Add Skykey Manager.
  - Add `siac skynet unpin` subcommand.
  -
  -